Add edit and delete buttons for admin users on recipe detail page

+++ recipes/views.py
@@ -97,6 +97,12 @@ def create_recipe(request):
 @login_required
 def edit_recipe(request, slug):
     recipe = get_object_or_404(Recipe, slug=slug)
+
+    # Check if the user is the author or an admin
+    if recipe.author != request.user and not request.user.is_staff:
+        messages.error(request, 'You do not have permission to edit this recipe.')
+        return redirect('recipe_detail', slug=slug)
+
     if request.method == 'POST':
         form = RecipeForm(request.POST, request.FILES, instance=recipe)
         if form.is_valid():
@@ -108,17 +114,32 @@ def edit_recipe(request, slug):
     else:
         form = RecipeForm(instance=recipe)
 
-    return render(request, 'recipes/edit_recipe.html', {'form': form, 'recipe': recipe, 'title': 'Edit Recipe'})
+    return render(request, 'recipes/edit_recipe.html', {
+        'form': form,
+        'recipe': recipe,
+        'title': 'Edit Recipe',
+    })
 
 # Recipe delete view
 @login_required
 def delete_recipe(request, slug):
     recipe = get_object_or_404(Recipe, slug=slug)
+
+    # Check if the user is the author or an admin
+    if recipe.author != request.user and not request.user.is_staff:
+        messages.error(request, 'You do not have permission to delete this recipe.')
+        return redirect('recipe_detail', slug=slug)
+
     if request.method == 'POST':
         recipe.delete()
         messages.success(request, 'Recipe deleted successfully!')
         return redirect('recipe_list')
-    return render(request, 'recipes/delete_recipe.html', {'recipe': recipe, 'title': 'Delete Recipe'})    
+
+    return render(request, 'recipes/delete_recipe.html', {
+        'recipe': recipe,
+        'title': 'Delete Recipe',
+    })
+
 
 # Custom login view using allauth
 class LoginView(AllauthLoginView):
